+++ packages/widgetbook/lib/src/next/addons/text_scale_addon.dart
@@ -1,8 +1,3 @@
-// [MediaQuery.textScaleFactor] is deprecated in Flutter 3.16.0,
-// Since our minimum Flutter version is 3.7.0, we cannot use [TextScaler] yet.
-// More info: https://docs.flutter.dev/release/breaking-changes/deprecate-textscalefactor
-// ignore_for_file: deprecated_member_use
-
 import 'package:flutter/widgets.dart';
 
 import '../../fields/fields.dart';
@@ -16,7 +11,7 @@ class TextScaleMode extends Mode<double> {
   Widget build(BuildContext context, Widget child) {
     return MediaQuery(
       data: MediaQuery.of(context).copyWith(
-        textScaleFactor: value,
+        textScaler: TextScaler.linear(value),
       ),
       child: child,
     );